+++ lib/src/validators/validators.dart
@@ -5,14 +5,16 @@
 import 'package:reactive_forms/reactive_forms.dart';
 import 'package:reactive_forms/src/validators/compose_validator.dart';
 import 'package:reactive_forms/src/validators/credit_card_validator.dart';
-
-import './email_validator.dart';
-import './max_length_validator.dart';
-import './min_length_validator.dart';
-import './must_match_validator.dart';
-import './number_validator.dart';
-import './pattern_validator.dart';
-import './required_validator.dart';
+import 'package:reactive_forms/src/validators/email_validator.dart';
+import 'package:reactive_forms/src/validators/equals_validator.dart';
+import 'package:reactive_forms/src/validators/max_length_validator.dart';
+import 'package:reactive_forms/src/validators/max_validator.dart';
+import 'package:reactive_forms/src/validators/min_length_validator.dart';
+import 'package:reactive_forms/src/validators/min_validator.dart';
+import 'package:reactive_forms/src/validators/must_match_validator.dart';
+import 'package:reactive_forms/src/validators/number_validator.dart';
+import 'package:reactive_forms/src/validators/pattern_validator.dart';
+import 'package:reactive_forms/src/validators/required_validator.dart';
 
 /// Signature of a function that receives a control and synchronously
 /// returns a map of validation errors if present, otherwise null.
@@ -29,6 +31,10 @@ class Validators {
   /// Gets a validator that requires the control have a non-empty value.
   static ValidatorFunction get required => RequiredValidator().validate;
 
+  /// Gets a validator that requires the control's value be true.
+  /// This validator is commonly used for required checkboxes.
+  static ValidatorFunction get requiredTrue => EqualsValidator(true).validate;
+
   /// Gets a validator that requires the control's value pass an email
   /// validation test.
   static ValidatorFunction get email => EmailValidator().validate;
@@ -39,6 +45,18 @@ class Validators {
   /// Gets a validator that validates if the control's value is a valid
   /// credit card number.
   static ValidatorFunction get creditCard => CreditCardValidator().validate;
+
+  /// Gets a validators that requires the control's value to be equals to
+  /// argument [value].
+  static ValidatorFunction equals<T>(T value) => EqualsValidator(T).validate;
+
+  /// Gets a validator that requires the control's value to be greater than
+  /// or equal to [min] value.
+  static ValidatorFunction min(Comparable min) => MinValidator(min).validate;
+
+  /// Gets a validator that requires the control's value to be less than
+  /// or equal to [max] value.
+  static ValidatorFunction max(Comparable max) => MaxValidator(max).validate;
 
   /// Gets a validator that requires the length of the control's value to be
   /// greater than or equal to the provided [minLength].